.claude/skills/kb/scripts/requirements.txt
PyMuPDF>=1.24
openpyxl>=3.1
pandas>=2.0
pytesseract>=0.3
python-docx>=1.1
Pillow>=10.0
